% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{metaHD-package}
\alias{metaHD}
\alias{metaHD-package}
\alias{metahd}
\title{correlateR}
\description{
This R-package correlateR features fast, robust, and efficient marginal,
partial, semi-partial correlations and covariances of arbitrary conditional
order.
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

